+++ python/cuml/linear_model/logistic_regression.pyx
@@ -27,8 +27,7 @@ import rmm
 from cuml.solvers import QN
 from cuml.common.base import Base
 from cuml.metrics.accuracy import accuracy_score
-from cuml.utils import input_to_dev_array
-from cuml.utils.cupy_utils import rmm_cupy_ary
+from cuml.utils import input_to_dev_array, rmm_cupy_ary
 
 
 supported_penalties = ['l1', 'l2', 'none', 'elasticnet']
